Make ToggleSelector its component
